rename all stts to sttList
